vigia_detect/requirements.txt
# LPP-Detect Requirements

# Básicas
numpy>=1.20.0
pillow>=8.0.0
opencv-python>=4.5.0
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# YOLOv5
torch>=1.7.0
torchvision>=0.8.0

# Supabase
supabase>=0.5.8
postgrest>=0.10.0

# Utilidades
matplotlib>=3.4.0
pandas>=1.3.0

# Para desarrollo
pytest>=6.0.0
pylint>=2.8.0

# ADK para agentes
google-cloud-aiplatform>=1.0.0

# Twilio y WhatsApp
twilio>=7.0.0
flask>=2.0.0
requests>=2.25.0

# Web (para fases posteriores)
fastapi>=0.70.0
uvicorn>=0.15.0
python-multipart>=0.0.5

# Redis and Vector Search
redis>=5.0.0
redisvl>=0.2.0
sentence-transformers>=2.2.0
transformers>=4.30.0
PyPDF2>=3.0.0
